a lot new stuff

+++ src/com/nkcoding/communication/Communication.java
@@ -36,7 +36,7 @@ public abstract class Communication implements Closeable {
      * @param peer         the id got from openCommunication
      * @param transmission the transmission to send
      */
-    public abstract void sendTo(int peer, ResetDataOutputStream transmission);
+    public abstract void sendTo(short peer, ResetDataOutputStream transmission);
 
     /**
      * sends some data to all peers
@@ -64,7 +64,7 @@ public abstract class Communication implements Closeable {
      *
      * @return a list with all peers
      */
-    public abstract Set<Integer> getPeers();
+    public abstract Set<Short> getPeers();
 
     public boolean isServer() {
         return isServer;
@@ -75,5 +75,5 @@ public abstract class Communication implements Closeable {
      *
      * @return the id
      */
-    public abstract int getId();
+    public abstract short getId();
 }
